Fix de bugs dans variableDelayMs et timer1 + Grille dans le main + ajout de fonctions dans line follower et wheels

+++ projet/lib/general_functions.cpp
@@ -2,7 +2,7 @@
 
 void variableDelayMs(uint16_t delay)
 {
-    for (uint8_t i = 0; i < (delay / 10); i++)
+    for (uint16_t i = 0; i < (delay / 10); i++)
     {
         _delay_ms(DELAY_MS_10);
     }
